+++ lib/src/bin/commands/commands.dart
@@ -1,3 +1,5 @@
+export 'aab_build_command.dart';
+export 'apk_build_command.dart';
 export 'build_command.dart';
 export 'encrypt_command.dart';
 export 'decrypt_command.dart';